updated close icon, created an extra file for new icon
